Create Hourglass.java
Write a Java program for a given 6x6 array having an hourglass. There can be many hourglass in the array. The sum of an hourglass is the sum of all the numbers within it. In this problem you have to print the largest sum among all the hourglasses in the array.